Constraint logic + Collisions
Two new class where added to handle possible constraint: Constraint and RewardManager. The former is an abstract class to define constraints, the latter is used to add/remove constraint and compute penalties. The collision constraint was created, together with an obstacle layer for the map. Furthermore the reward for advancing on the path was made more flexible.